recuirter/requirements.txt
python-dotenv==1.0.0
PyPDF2==3.0.1
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.97.0
python-dateutil==2.8.2
